unify unit test files
